requirements.txt
Django
djangorestframework
pillow
psycopg2-binary
python-dotenv
drf-yasg
twilio